requirements.txt
tensorflow>=1.12.1
absl-py==0.5.0
altgraph==0.16.1
astor==0.7.1
future==0.16.0
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
macholib==1.11
Markdown==3.0.1
numpy==1.14.5
pefile==2018.8.8
Pillow==5.3.0
protobuf==3.6.1
pywin32-ctypes==0.2.0
PyYAML==4.2b1
scipy==1.1.0
six==1.11.0
tensorboard==1.10.0
termcolor==1.1.0
Werkzeug==0.14.1
